modularization
divide drive file with lane detection && traffic light && publishing drive
